Get project progress

Fetch the progress of a ClimMob project

getProjectProgress(key, project, server = "climmob3", ...)

Arguments

key
a character for the user's application programming interface (API) key

project
a character with the id of one or more projects

server
optional, a character to select from which server the data will be retrieved. See details

...
additional arguments passed to methods. See details

Value

A data frame with the ClimMob projects

project_id

the project unique id

name

the project name

moment

either the design, registration or data collection

number_obs

number of observations collected in a given moment

last_activity

last activity of the given moment

Details

server: the default server is "climmob" used for clients of https://climmob.net/climmob3/, other options are:

"avisa" for clients of https://avisa.climmob.net/

"rtb" for clients of https://rtb.climmob.net/

"testing" for clients of https://testing.climmob.net/climmob3/

See also

Author

Kauê de Sousa

Examples

if (FALSE) { 
# This function will not work without an API key  
# the user API key can be obtained once a free ClimMob account 
# is created via https://climmob.net/

my_key <- "add_your_key"

my_project <- "project_id"

getProjectProgress(my_key, my_project)

}