Remove geographical identity

Build a buffer around the a set of geographical coordinates and take a random point around the buffer. The function is used to omit the precise location of tricot participants but keeping a close distance to its agro-environment

rmGeoIdentity(lonlat, dist = 0.015, nQuadSegs = 2L, ...)

Arguments

lonlat
a data.frame or matrix with geographical coordinates long lat

dist
numeric, buffer distance for all lonlat

nQuadSegs
integer, number of segments per quadrant

...
further arguments passed to sf methods

Examples

xy <- matrix(c(11.097799, 60.801090,
               11.161298, 60.804199,
               11.254428, 60.822457),
             nrow = 3, ncol = 2, byrow = TRUE)

rmGeoIdentity(xy)
#>          x        y
#> 1 11.10416 60.80445
#> 2 11.16589 60.80275
#> 3 11.25493 60.82776

# the function handles NAs by keeping then 
# in a logic vector to reconstruct the matrix
xy2 <- matrix(c(11.097799, 60.801090,
                NA, NA,
                11.161298, 60.804199,
                11.254428, 60.822457,
                11.254428, NA),
              nrow = 5, ncol = 2, byrow = TRUE)

rmGeoIdentity(xy2)
#>          x        y
#> 1 11.10923 60.80803
#> 2       NA       NA
#> 3 11.15826 60.80211
#> 4 11.24379 60.82099
#> 5       NA       NA