Build Plackett-Luce rankings

Create an object of class "rankings" or "grouped_rankings" from tricot data

build_rankings(data = NULL, items = NULL, input = NULL,
  additional.rank = NULL, grouped.rankings = FALSE, ...)

Arguments

data
a data frame with columns specified by items and input values to rank.

items
a data frame or index of data for the column(s) containing the item names

input
a data frame or index of data for the column(s) containing the values to be ranked

additional.rank
optional, a data frame for the comparisons between tricot items and the local item

grouped.rankings
optional, if TRUE return an object of class grouped_rankings

...
additional arguments passed to methods

Value

a PlackettLuce "rankings" object, which is a matrix of dense rankings

See also

Examples

# beans data where each observer compares 3 varieties randomly distributed 
# from a list of 11 and additionally compares these 3 varieties 
# with their local variety
library("PlackettLuce")
data("beans", package = "PlackettLuce")

# first build rankings with only tricot items
# and return an object of class 'rankings'
R <- build_rankings(data = beans,
                    items = c(1:3),
                    input = c(4:5))
head(R)

#> [1] "PM2 Don Rey > SJC 730-79 > BRT 103-182"  
#> [2] "INTA Centro Sur > INTA Sequia > INTA ..."
#> [3] "INTA Ferroso > INTA Matagalpa > BRT  ..."
#> [4] "INTA Rojo > INTA Centro Sur > ALS 0532-6"
#> [5] "PM2 Don Rey > INTA Sequia > SJC 730-79"  
#> [6] "ALS 0532-6 > INTA Matagalpa > INTA Rojo" 

############################################################

# pass the comparison with local item as an additional rankings, then
# each of the 3 varieties are compared separately with the local item
# and return an object of class grouped_rankings
G <- build_rankings(data = beans,
                    items = c(1:3),
                    input = c(4:5),
                    additional.rank = beans[c(6:8)],
                    grouped.rankings = TRUE)

head(G)

#>                                                 1 
#>  "PM2 Don Rey > SJ ..., Local > BRT 103-182, ..." 
#>                                                 2 
#>    "INTA Centro Sur  ..., Local > INTA Rojo, ..." 
#>                                                 3 
#> "INTA Ferroso > I ..., INTA Ferroso > Local, ..." 
#>                                                 4 
#> "INTA Rojo > INTA ..., INTA Centro Sur  ..., ..." 
#>                                                 5 
#>  "PM2 Don Rey > IN ..., INTA Sequia > Local, ..." 
#>                                                 6 
#>   "ALS 0532-6 > INT ..., ALS 0532-6 > Local, ..." 

############################################################

# rankings with more than 3 items in a tricot project
# 5 items randomly assigned

input <- as.data.frame(matrix(NA, nrow = 5, ncol = 5))
names(input) <- paste0("position_item_",LETTERS[1:5])

for(s in 1:5) {
  input[s,] <- sample(1:5)
}

items <- randomise(5, 5, 5, c("green","blue","red","white","yellow"))

build_rankings(items = items,
               input = input)

#>                                     1                                     2 
#> "green > red > yellow > blue > white" "white > green > blue > yellow > red" 
#>                                     3                                     4 
#> "white > blue > green > yellow > red" "white > yellow > red > green > blue" 
#>                                     5 
#> "blue > yellow > red > green > white"